Bump, translations, reorder
