+++ core-ui/src/components/Predefined/Details/Secret/ChartContent.js
@@ -1,0 +1,36 @@
+import React, { useState } from 'react';
+import { ReadonlyEditorPanel } from 'shared/components/ReadonlyEditorPanel';
+import { ComboboxInput } from 'shared/ResourceForm/inputs';
+import { useTranslation } from 'react-i18next';
+
+export function ChartContent({ chart }) {
+  const { t } = useTranslation();
+
+  const files = [...(chart?.files || []), ...(chart?.templates || [])];
+
+  const options = files.map(({ name, data }) => ({
+    text: name,
+    key: name,
+    data,
+  }));
+
+  const [currentFile, setCurrentFile] = useState(options[0]);
+
+  const actions = (
+    <div style={{ width: '300px' }}>
+      <ComboboxInput
+        value={currentFile?.key}
+        options={options}
+        onSelectionChange={(_, selected) => setCurrentFile(selected)}
+      />
+    </div>
+  );
+
+  return (
+    <ReadonlyEditorPanel
+      title={t('secrets.helm.chart-files')}
+      value={atob(currentFile?.data || '')}
+      actions={actions}
+    />
+  );
+}
