+++ core-ui/src/components/Predefined/Details/Secret/HelmReleaseData.js
@@ -1,40 +1,14 @@
-import React, { useState } from 'react';
+import React from 'react';
 import { decodeHelmRelease } from './decodeHelmRelease';
 import { ReadonlyEditorPanel } from 'shared/components/ReadonlyEditorPanel';
 import { ReleaseData } from './ReleaseData';
-import { ComboboxInput } from 'shared/ResourceForm/inputs';
-
-function ChartContent({ chart }) {
-  const files = [...(chart?.files || []), ...(chart?.templates || [])];
-
-  const options = files.map(({ name, data }) => ({
-    text: name,
-    key: name,
-    data,
-  }));
-
-  const [currentFile, setCurrentFile] = useState(options[0]);
-
-  const actions = (
-    <div style={{ width: '300px' }}>
-      <ComboboxInput
-        value={currentFile?.key}
-        options={options}
-        onSelectionChange={(_, selected) => setCurrentFile(selected)}
-      />
-    </div>
-  );
-
-  return (
-    <ReadonlyEditorPanel
-      title="Chart Files"
-      value={atob(currentFile?.data || '')}
-      actions={actions}
-    />
-  );
-}
+import { ChartContent } from './ChartContent';
+import { useTranslation } from 'react-i18next';
+import jsyaml from 'js-yaml';
 
 export function HelmReleaseData(secret) {
+  const { t } = useTranslation();
+
   if (secret.type !== 'helm.sh/release.v1') {
     return null;
   }
@@ -47,8 +21,8 @@ export function HelmReleaseData(secret) {
     <React.Fragment key="helm-release-data">
       <ReleaseData release={release} />
       <ReadonlyEditorPanel
-        title="Release config"
-        value={JSON.stringify(release.config, null, 2)}
+        title={t('secrets.helm.release-config')}
+        value={jsyaml.dump(release.config)}
       />
       <ChartContent chart={release.chart} />
     </React.Fragment>
