+++ core/src/i18n/en.yaml
@@ -1665,6 +1665,16 @@ secrets:
       success: Secret updated.
   headers:
     type: Type
+  helm:
+    chart-description: Chart Description
+    chart-files: Chart Files
+    chart-name: Chart Name
+    chart-version: Chart Version
+    data-title: Release {{name}}
+    first-deployed: First Deployed
+    last-deployed: Last Deployed
+    release-config: Release config
+    release-version: Version v{{version}}
   labels:
     annotations: Annotations
     type: Secret Type
